feat: [CDS-58493]: Sync step - change strategy from sync to async for log streaming (#46036)
* feat: [CDS-58493]: Sync step - change strategy from sync to async for log streaming (#45976)

* testing fixes

* Merge remote-tracking branch 'origin/develop' into cds-54817

* fixes

* Merge remote-tracking branch 'origin/develop' into cds-54817

* add envs

* fix format

* make appslist as runtime

* change str

* fix failing test

* fix failing test

* fix failing test

* feat: [CDS-58493]: Sync step - change strategy from sync to async for log streaming

* fix conflicts

* fix failing test

* Sync step - add null check for toBool and toNumber
