+++ 125-cd-nextgen/src/main/java/io/harness/cdng/gitops/syncstep/SyncRetryStrategy.java
@@ -16,7 +16,6 @@ import io.harness.yaml.YamlSchemaTypes;
 
 import io.swagger.annotations.ApiModelProperty;
 import javax.validation.constraints.Min;
-import javax.validation.constraints.NotNull;
 import lombok.Builder;
 import lombok.Data;
 
@@ -27,20 +26,14 @@ public class SyncRetryStrategy {
   @YamlSchemaTypes(value = {string})
   @ApiModelProperty(dataType = SwaggerConstants.INTEGER_CLASSPATH)
   @Min(0)
-  @NotNull
   public ParameterField<Integer> limit;
 
-  @ApiModelProperty(dataType = SwaggerConstants.STRING_CLASSPATH)
-  @NotNull
-  public ParameterField<String> baseBackoffDuration;
+  @ApiModelProperty(dataType = SwaggerConstants.STRING_CLASSPATH) public ParameterField<String> baseBackoffDuration;
 
   @YamlSchemaTypes(value = {string})
   @ApiModelProperty(dataType = SwaggerConstants.INTEGER_CLASSPATH)
   @Min(0)
-  @NotNull
   public ParameterField<Integer> increaseBackoffByFactor;
 
-  @ApiModelProperty(dataType = SwaggerConstants.STRING_CLASSPATH)
-  @NotNull
-  public ParameterField<String> maxBackoffDuration;
+  @ApiModelProperty(dataType = SwaggerConstants.STRING_CLASSPATH) public ParameterField<String> maxBackoffDuration;
 }
