+++ 125-cd-nextgen/src/main/java/io/harness/cdng/gitops/syncstep/SyncStepInfo.java
@@ -45,7 +45,7 @@ public class SyncStepInfo extends SyncBaseStepInfo implements CDAbstractStepInfo
 
   @Override
   public String getFacilitatorType() {
-    return OrchestrationFacilitatorType.SYNC;
+    return OrchestrationFacilitatorType.ASYNC;
   }
 
   @Override
