+++ 125-cd-nextgen/src/main/java/io/harness/serializer/kryo/NGKryoRegistrar.java
@@ -42,6 +42,7 @@ import io.harness.cdng.gitops.UpdateReleaseRepoStepInfo;
 import io.harness.cdng.gitops.UpdateReleaseRepoStepParams;
 import io.harness.cdng.gitops.beans.FetchLinkedAppsStepParams;
 import io.harness.cdng.gitops.beans.GitOpsLinkedAppsOutcome;
+import io.harness.cdng.gitops.syncstep.SyncResponse;
 import io.harness.cdng.gitops.syncstep.SyncStepOutcome;
 import io.harness.cdng.helm.HelmDeployStepInfo;
 import io.harness.cdng.helm.HelmDeployStepParams;
@@ -136,6 +137,7 @@ import io.harness.cdng.tas.TasSwapRoutesStepInfo;
 import io.harness.cdng.tas.TasSwapRoutesStepParameters;
 import io.harness.cdng.tasks.manifestFetch.step.ManifestFetchOutcome;
 import io.harness.cdng.tasks.manifestFetch.step.ManifestFetchParameters;
+import io.harness.gitops.models.Application;
 import io.harness.serializer.KryoRegistrar;
 import io.harness.telemetry.beans.CdTelemetrySentStatus;
 
@@ -290,5 +292,7 @@ public class NGKryoRegistrar implements KryoRegistrar {
     kryo.register(SyncStepOutcome.class, 12665);
     kryo.register(TasRouteMappingStepInfo.class, 12666);
     kryo.register(TasRouteMappingStepParameters.class, 12667);
+    kryo.register(SyncResponse.class, 12668);
+    kryo.register(Application.class, 12669);
   }
 }
