+++ 125-cd-nextgen/src/test/resources/schema/DeploymentStage/all.json
@@ -15098,12 +15098,6 @@
     },
     "SyncRetryStrategy": {
       "type": "object",
-      "required": [
-        "baseBackoffDuration",
-        "increaseBackoffByFactor",
-        "limit",
-        "maxBackoffDuration"
-      ],
       "properties": {
         "baseBackoffDuration": {
           "type": "string"
